file creation
log_core : uses a template with variables for easy printing. Want to add it to a text file in the future

log_db: database creation. Uses the explorer to manually select a .xlsx or .xls file, then converts it into a .db file, pretty cool

notes: must replace empty cells with null or something, gotta remove that